Improve Alder Lake thread. Fix display errors.

+++ algo/gr/gr-gate.h
@@ -228,7 +228,7 @@ void tune(void *input, int thr_id);
 
 void benchmark(void *input, int thr_id, long sleep_time);
 void stress_test(void *input, int thr_id);
-bool is_thread_used(int thr_id);
+bool is_thread_used(size_t thr_id);
 size_t get_config_id();
 bool check_prepared();
